ReClean resolution handling (#349)
This needed a bit more than my initial optimistic commit.

It had an argument named `--fmri-resolution` which appears to have
actually needed to match the grayordinate resolution. So, I have done
some renaming in addition to exposing the extra arguments.

It should still be compatible with the previous Batch (compatibility
flag name, defaults for the added options).

It has not been tested.

---------

Co-authored-by: Jure Demsar <demsarjure@gmail.com>